+++ R/xgb.R
@@ -58,23 +58,6 @@ axe_call.xgb.Booster <- function(x, verbose = FALSE, ...) {
   )
 }
 
-#' Remove controls used for training.
-#'
-#' @rdname axe-xgb.Booster
-#' @export
-axe_ctrl.xgb.Booster <- function(x, verbose = FALSE, ...) {
-  old <- x
-  x <- exchange(x, "params", list(NULL))
-
-  add_butcher_attributes(
-    x,
-    old,
-    disabled = c("print()"),
-    add_class = FALSE,
-    verbose = verbose
-  )
-}
-
 #' Remove environments.
 #'
 #' @rdname axe-xgb.Booster
@@ -89,24 +72,6 @@ axe_env.xgb.Booster <- function(x, verbose = FALSE, ...) {
   add_butcher_attributes(
     x,
     old,
-    add_class = FALSE,
-    verbose = verbose
-  )
-}
-
-#' Remove cached memory dump of xgboost model that was saved as a raw
-#' data type.
-#'
-#' @rdname axe-xgb.Booster
-#' @export
-axe_fitted.xgb.Booster <- function(x, verbose = FALSE, ...) {
-  old <- x
-  x <- exchange(x, "raw", raw())
-
-  add_butcher_attributes(
-    x,
-    old,
-    disabled = c("xgb.Booster.complete()"),
     add_class = FALSE,
     verbose = verbose
   )