+++ NEWS.md
@@ -5,6 +5,9 @@
 
 * Added an `axe_fitted()` method to butcher the `template` slot for prepped 
   recipes (@AshesITR, #207).
+  
+* Removed `axe_fitted()` and `axe_ctrl()` for xgboost, because these methods
+  caused problems for prediction (#218).
 
 # butcher 0.1.5